refactor: Move interface
