+++ core/common/src/main/java/com/teamoffroad/core/common/util/IntentProvider.kt
@@ -1,4 +1,4 @@
-package com.teamoffroad.core.common.intentProvider
+package com.teamoffroad.core.common.util
 
 import android.content.Intent
