+++ feature/main/src/main/java/com/teamoffroad/feature/main/di/IntentModule.kt
@@ -1,6 +1,6 @@
 package com.teamoffroad.feature.main.di
 
-import com.teamoffroad.core.common.intentProvider.IntentProvider
+import com.teamoffroad.core.common.util.IntentProvider
 import com.teamoffroad.feature.main.intentprovider.IntentProviderImpl
 import dagger.Binds
 import dagger.Module
